hello/pyproject.toml
[tool.poetry]
name = "hello"
version = "0.1.0"
description = ""
authors = ["Imran691 <mimran691@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
